# API/swagger.yaml
...
swagger: "2.0"
info:
  description: "Student Service API"
  version: "1.0.0"
  title: "Student"
  termsOfService: ""
  contact:
    name: "someName"
    url: "http://www.staffshift.com"
    email: "info@staffshift.com"
host: "localhost:3000"
basePath: "/"
...
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /admin/technology/get-technology:
    get:
      tags:
        - Admin
      summary: "get technology"
      x-swagger-router-controller: admin-swagger
      operationId: getAllTechnology
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: "return all technology"
          schema:
            $ref: "#/definitions/allTechnology_payload"
        400:
          $ref: "#/responses/admin_credentials_error"
        default:
          $ref: "#/responses/default_server_error"
...
definitions:
...
  server_error_model:
    type: object
    required:
      - message
    properties:
      message:
        type: string
      exception:
        $ref: "#/definitions/exception_model"
  exception_model:
    type: object
...
  common_validation_error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        enum:
          - PATTERN
          - REQUIRED
          - MIN_LENGTH
          - DATA_EXIST
          - DATA_NOT_EXIST
      message:
        type: string
...
  mongoose_default:
    type: object
    required:
      - _id
      - __v
    properties:
      _id:
        type: object
        pattern: ^[0-9a-fA-F]{24}$
      __v:
        type: number
...
  technology_payload:
    type: object
    required:
      - technology
      - hours
    properties:
      technology:
        type: string
        default: node js
        minLength: 5
      hours:
        type: number
        minLength: 2
  technology_response_payload:
    allOf:
      - $ref: "#/definitions/technology_payload"
      - $ref: "#/definitions/mongoose_default"
  allTechnology_payload:
    type: object
    properties:
      data:
        type: array
        items:
          - $ref: "#/definitions/technology_response_payload"
...
responses:
...
  default_server_error:
    description: "unexpected error"
    schema:
      $ref: "#/definitions/server_error_model"
...
  admin_credentials_error:
    description: "error on create student"
    schema:
      $ref: "#/definitions/common_validation_error"
...